+++ src/components/Positions/index.tsx
@@ -2,10 +2,10 @@ import { twMerge } from "tailwind-merge";
 
 import { usePositions } from "@/hooks/usePositions";
 
-import { LoadingDots } from "./LoadingDots";
-import { Position } from "./Position";
+import { LoadingDots } from "../LoadingDots";
+import { SinglePosition } from "./SinglePosition";
 import { PositionColumn } from "./PositionColumn";
-import { PoolTokens } from "./PoolTokens";
+import { PoolTokens } from "../PoolTokens";
 import { useWallet } from "@solana/wallet-adapter-react";
 
 interface Props {
@@ -58,7 +58,7 @@ export function Positions(props: Props) {
               <PositionColumn num={7}>Liq. Price</PositionColumn>
             </div>
             {pool.positions.map((position, i) => (
-              <Position
+              <SinglePosition
                 className={twMerge(
                   "border-zinc-700",
                   i < pool.positions.length - 1 && "border-b"
